Revamp trust_list.get_path() to not include TRUSTED CERTIFICATE format
From user reports, at least some versions of OpenSSL will fail to authenticate
a TLS connection server certificate if there is one or more TRUSTED CERTIFICATE
objects in the CA bundle.

The intent of this function has always been to export a CA bundle for use with
OpenSSL to make TLS connections. Thus, the map_vendor_oids parameter was
removed, and the function explicitly excludes any certificates that the OS
does not trust for TLS server authentication.