Allow re-calling use_ctypes(), use_openssl() and use_winlegacy()
Previously if these were called after any part of oscrypto had been
initialized it would be an error. Now as long as the configuration is
the same, they can be called multiple times. This can help ensure
config is consistent when there are different entry points into oscrypto
from an app that is configuring it.