Fix a bug with the OpenSSL backend and using AES192/256 and no padding
With AES192 it was not possible to encrypt data unless they had a length
that was a multiple of both 24 and 16 bytes.

With AES256, it would only allow encrypting plaintext that had a length
that was a multiple of 32 bytes.

Both should only restrict plaintext without padding have a length that
is a multiple of the block size, which is 16 bytes.